backend/security/requirements-security.txt
# Security dependencies for AtonixCorp Platform
# Add these to your requirements.txt

# Cryptography
cryptography>=41.0.0
PyJWT>=2.8.0

# Rate limiting and security
django-ratelimit>=4.1.0
django-axes>=6.1.0  # Login attempt tracking
django-cors-headers>=4.3.0

# Input validation and sanitization
bleach>=6.0.0  # HTML sanitization
python-decouple>=3.8  # Environment variables

# SSL/TLS and certificates
certifi>=2023.7.22
requests[security]>=2.31.0

# Security monitoring
django-security>=0.17.0
django-defender>=0.9.7  # Brute force protection

# Redis for caching and rate limiting
redis>=4.6.0
django-redis>=5.3.0

# Additional security tools
python-jose[cryptography]>=3.3.0  # JWT with cryptography
passlib[bcrypt]>=1.7.4  # Password hashing
argon2-cffi>=23.1.0  # Argon2 password hashing

# Environment and configuration
python-dotenv>=1.0.0

# Development security tools (optional)
bandit>=1.7.5  # Security linter
safety>=2.3.0  # Dependency vulnerability scanner